feat: adding ChatZone component
